# public/swagger/swagger.json
{"swagger":"2.0","info":{"title":"Scheduler Inventory API Docs","version":"1.0.0","description":"API for managing requisitions of nextcoming events"},"paths":{"/catalog":{"get":{"tags":["Products"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"start_date","in":"path","required":true,"type":"string","description":"Filter by start date"},{"name":"end_date","in":"path","required":true,"type":"string","description":"Filter by end date"},{"name":"token","in":"path","required":true,"type":"string","description":"User token"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/List"}}}}}},"/catalog/{id}":{"get":{"tags":["Products"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Product ID"},{"name":"start_date","in":"path","required":true,"type":"string","description":"Filter by start date"},{"name":"end_date","in":"path","required":true,"type":"string","description":"Filter by end date"},{"name":"token","in":"path","required":true,"type":"string","description":"User token"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Product"}}}}}},"/requisitions":{"post":{"tags":["Requisitions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/CreateRequisition"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Requisition"}}}}},"get":{"tags":["Requisitions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"page","in":"path","required":true,"type":"number","description":"Page number"},{"name":"limit","in":"path","required":true,"type":"number","description":"Number of items per page"},{"name":"search","in":"path","required":true,"type":"string","description":"Search by event name"},{"name":"approved","in":"path","required":true,"type":"boolean","description":"Filter by approval status"},{"name":"user_id","in":"path","required":true,"type":"string","description":"Filter by user ID"},{"name":"start_date","in":"path","required":true,"type":"date","description":"Filter by start date"},{"name":"end_date","in":"path","required":true,"type":"date","description":"Filter by end date"},{"name":"active","in":"path","required":true,"type":"boolean","description":"Filter by active status"},{"name":"token","in":"path","required":true,"type":"string","description":"User token"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/List"}}}}}},"/requisitions/{id}/delete":{"patch":{"tags":["Requisitions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Requisition ID"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/DeleteRequisition"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Requisition"}}}}}},"/requisitions/{id}":{"get":{"tags":["Requisitions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Requisition ID"},{"name":"token","in":"path","required":true,"type":"string","description":"User token"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Requisition"}}}}},"put":{"tags":["Requisitions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Requisition ID"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/UpdateRequisition"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Requisition"}}}}}},"/requisitions/{id}/restore":{"patch":{"tags":["Requisitions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Requisition ID"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/RestoreRequisition"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Requisition"}}}}}},"/requisitions/{id}/review":{"patch":{"tags":["Requisitions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Requisition ID"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/ReviewRequisition"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Requisition"}}}}}}},"definitions":{"GetCatalog":{"properties":{"start_date":{"type":"string","description":"Filter by start date"},"end_date":{"type":"string","description":"Filter by end date"},"token":{"type":"string","description":"User token"}},"required":["start_date","end_date","token"]},"List":{"properties":{},"required":[]},"GetProductById":{"properties":{"id":{"type":"string","description":"Product ID"},"start_date":{"type":"string","description":"Filter by start date"},"end_date":{"type":"string","description":"Filter by end date"},"token":{"type":"string","description":"User token"}},"required":["id","start_date","end_date","token"]},"Product":{"properties":{},"required":[]},"CreateRequisition":{"properties":{"event_name":{"type":"string","description":"Event name"},"start_date":{"type":"date","description":"Start date of the event"},"end_date":{"type":"date","description":"End date of the event"},"required_products":{"type":"object","description":"Required products","$ref":"#/definitions/required_products"},"address":{"type":"string","description":"Address"},"token":{"type":"string","description":"User token"}},"required":["event_name","start_date","end_date","required_products","address","token"]},"Requisition":{"properties":{},"required":[]},"GetRequisitions":{"properties":{"page":{"type":"number","description":"Page number"},"limit":{"type":"number","description":"Number of items per page"},"search":{"type":"string","description":"Search by event name"},"approved":{"type":"boolean","description":"Filter by approval status"},"user_id":{"type":"string","description":"Filter by user ID"},"start_date":{"type":"date","description":"Filter by start date"},"end_date":{"type":"date","description":"Filter by end date"},"active":{"type":"boolean","description":"Filter by active status"},"token":{"type":"string","description":"User token"}},"required":["page","limit","search","approved","user_id","start_date","end_date","active","token"]},"DeleteRequisition":{"properties":{"id":{"type":"string","description":"Requisition ID"},"token":{"type":"string","description":"User token"}},"required":["id","token"]},"GetRequisitionById":{"properties":{"id":{"type":"string","description":"Requisition ID"},"token":{"type":"string","description":"User token"}},"required":["id","token"]},"UpdateRequisition":{"properties":{"id":{"type":"string","description":"Requisition ID"},"event_name":{"type":"string","description":"Event name"},"start_date":{"type":"date","description":"Start date of the event"},"end_date":{"type":"date","description":"End date of the event"},"required_products":{"type":"object","description":"Required products","$ref":"#/definitions/required_products"},"address":{"type":"string","description":"Address"},"token":{"type":"string","description":"User token"}},"required":["id","event_name","start_date","end_date","required_products","address","token"]},"RestoreRequisition":{"properties":{"id":{"type":"string","description":"Requisition ID"},"token":{"type":"string","description":"User token"}},"required":["id","token"]},"ReviewRequisition":{"properties":{"id":{"type":"string","description":"Requisition ID"},"review":{"type":"string","description":"Review text"},"token":{"type":"string","description":"User token"}},"required":["id","review","token"]}}}